Added cooldowns for items. Cooldowns are shown on hotkey buttons like spells, and on the examine_item screen. Cooldowns are stored at the creature level, so if a creature drops an item on cooldown and another one picks it up, cooldown won't exist for the second. Made Item:target() and Item:use() more like the same functions in Spell inventory.lua now doesn't assume it's the player's inventory, can pass in any creature (untested, and examine_item.lua is not updated) player is now forceMove()'d at the beginning of the game rather than moveTo()'d, hopefully to prevent a crash that I think was being caused by a creature being where the player is supposed to spawn

+++ data/items.lua
@@ -255,8 +255,12 @@ end
 function scroll:use(target,user)
   if possibleSpells[self.spell].target_type == "self" then
     target = user
-    possibleSpells[self.spell]:use(target,user,true,true)
-    user:delete_item(self)
+    local spResult = possibleSpells[self.spell]:use(target,user,true,true)
+    if spResult ~= false then
+      user:delete_item(self)
+    else
+      return false
+    end
   else
     action="targeting"
     actionResult=possibleSpells[self.spell]
@@ -706,7 +710,7 @@ function painwand:use(target,user)
     local dmg = target:damage(5)
     if player:can_sense_creature(target) then output:out(user:get_name() .. " blasts " .. target:get_name() .. " with a wand, dealing " .. dmg .. " damage.") end
     self.charges = self.charges-1
-    self.value = math.max(self.charges*10,1)
+    self.value = math.max(self.charges*5,1)
     return true
   else
     return false
